[All] Simplify ms2time, use all uint32_t
git-svn-id: svn://svn.berlios.de/avidemux/branches/avidemux_2.6_branch_mean@5434 30592fff-cfd4-0310-bf0e-8235a3c7be26

+++ avidemux/common/ADM_commonUI/DIA_gototime.cpp
@@ -2,7 +2,7 @@
 #include "ADM_default.h"
 #include "DIA_factory.h"
 
-uint8_t DIA_gotoTime(uint16_t *hh, uint16_t *mm, uint16_t *ss)
+uint8_t DIA_gotoTime(uint32_t *hh, uint32_t *mm, uint32_t *ss)
 {
 uint32_t h=*hh,m=*mm,s=*ss;
